- make the function as a part of commcare activity
